Move luma_ac to predict module
